Create a classification_result instance

Create a classification_result instance

classification_result(observed, predicted, probs, testind = NULL,
  test_design = NULL, predictor = NULL)

binary_classification_result(observed, predicted, probs, testind = NULL,
  test_design = NULL, predictor = NULL)

multiway_classification_result(observed, predicted, probs,
  testind = NULL, test_design = NULL, predictor = NULL)

regression_result(observed, predicted, testind = NULL,
  test_design = NULL, predictor = NULL)

Arguments

observed
a vector of observed/true values.

predicted
a vector of the predicted values.

probs
a matrix of predicted probabilities, one column per level.

testind
the row indices of the test observations.

test_design
an optional design for the test data.

predictor
an optional predictor object.

Examples

## a vector of observed values
yobs <- factor(rep(letters[1:4], 5))

## predicted probabilities
probs <- data.frame(a=runif(1:20), b=runif(1:20), c=runif(1:20), d=runif(1:20))
probs <- sweep(probs, 1, rowSums(probs), "/")

## get the max probability per row and use this to determine the predicted class
maxcol <- max.col(probs)
predicted <- levels(yobs)[maxcol]

## construct a classification result
cres <- classification_result(yobs, predicted, probs)

## compute default performance measures (Accuracy, AUC)
performance(cres)

#>    Accuracy         AUC 
#>  0.20000000 -0.02666667