Region of interest based MVPA analysis

Run a separate MVPA analysis for multiple disjoint regions of interest.

run_regional(
  model_spec,
  region_mask,
  return_fits = FALSE,
  compute_performance = TRUE,
  coalesce_design_vars = FALSE
)

Arguments

model_spec
a mvpa_model instance

region_mask
a NeuroVol or NeuroSurface where each region is identified by a unique integer. Every non-zero set of positive integers will be used to define a set of voxels for clasisifcation analysis.

return_fits
whether to return model fit for every ROI (default is FALSE to save memory)

compute_performance
logical indicating whether to compute performance measures (e.g. Accuracy, AUC)

Value

a named list of NeuroVol objects, where each element contains performance metric and is labelled according to the metric used (e.g. Accuracy, AUC)

a list of type regional_mvpa_result