Default method for run_regional

By default, if an object does not implement a specific run_regional.* method, this fallback will attempt to run the run_regional_base pipeline.

Usage

# Default S3 method
run_regional(model_spec, region_mask, ...)

Arguments

model_spec

A generic model specification object.

region_mask

A mask representing different brain regions.

...

Additional arguments passed to run_regional_base.

Value

A regional_mvpa_result object.