select_features

Given a feature_selection specification object and a dataset return the set of selected features as a binary vector.

select_features(obj, X, Y, ...)

# S3 method for catscore
select_features(obj, X, Y,
  ranking.score = c("entropy", "avg", "max"), ...)

# S3 method for FTest
select_features(obj, X, Y, ...)

Arguments

obj
the feature_selection object

X
region of interest containing training features, a class of type ROIVolume or ROISurface or matrix

Y
the dependent variable as a factor or numeric variable.

...
additional arguments

ranking.score
the feature score: entropy, avg, or max.

Value

a logical vector indicating the columns of X matrix that were selected.

Examples

fsel <- feature_selector("FTest", "top_k", 2)
coords <- rbind(c(1,1,1), c(2,2,2), c(3,3,3))

ROI <- neuroim2::ROIVec(neuroim2::NeuroSpace(c(10,10,10)), coords=coords, matrix(rnorm(100*3), 100, 3))

#> Found more than one class "AxisSet3D" in cache; using the first, from namespace 'neuroim'

#> Also defined by ‘neuroim2’

#> Found more than one class "AxisSet3D" in cache; using the first, from namespace 'neuroim'

#> Also defined by ‘neuroim2’

Y <- factor(rep(c("a", "b"), each=50))
featureMask <- select_features(fsel, ROI@data, Y)

#> selecting features via FTest

#> cutoff type top_k

#> cutoff value 2

#> Error in as.matrix(x): no slot of name "data" for this object of class "ROIVec"

sum(featureMask) == 2

#> Error in eval(expr, envir, enclos): object 'featureMask' not found

fsel2 <- feature_selector("FTest", "top_p", .1)
featureMask <- select_features(fsel2, ROI@data, Y)

#> selecting features via FTest

#> cutoff type top_p

#> cutoff value 0.1

#> Error in as.matrix(x): no slot of name "data" for this object of class "ROIVec"