Train an MVPA Model

This function trains a Multi-Variate Pattern Analysis (MVPA) model on the provided data, taking care of feature selection, parameter tuning, and model fitting.

Usage

# S3 method for class 'mvpa_model'
train_model(obj, train_dat, y, indices, wts = NULL, ...)

Arguments

obj

An object of class mvpa_model, specifying the MVPA problem.

train_dat

Training data, an instance of class ROIVolume or ROISurface.

y

The dependent variable (response variable), either a numeric vector or a factor.

indices

The spatial indices associated with each column.

wts

Optional class weights (if the underlying model supports it).

...

Additional arguments passed to other methods.

Value

A model fit object containing the trained model, its fit, the model type (classification or regression), the best tuning parameters, the voxel indices, and the feature mask.