Create an MVPA Dataset Object

Creates a dataset object for MVPA analysis that encapsulates a training dataset, an optional test dataset, and a voxel mask.

Usage

mvpa_dataset(train_data, test_data = NULL, mask)

Arguments

train_data

The training data set: a NeuroVec instance

test_data

Optional test data set: a NeuroVec instance (default: NULL)

mask

The set of voxels to include: a NeuroVol instance

Value

An mvpa_dataset object (S3 class) containing:

train_data

The training data as a NeuroVec instance

test_data

The test data as a NeuroVec instance (if provided, otherwise NULL)

mask

The binary mask defining valid voxels as a NeuroVol instance

has_test_set

Logical flag indicating whether this dataset has a test set

See also

Examples

# Create dataset from NeuroVec objects
train_vec <- NeuroVec(array(rnorm(1000*100), c(10,10,10,100)))
#> Error in NeuroVec(array(rnorm(1000 * 100), c(10, 10, 10, 100))): could not find function "NeuroVec"
mask_vol <- NeuroVol(array(1, c(10,10,10)))
#> Error in NeuroVol(array(1, c(10, 10, 10))): could not find function "NeuroVol"
dataset <- mvpa_dataset(train_vec, mask=mask_vol)
#> Error: object 'train_vec' not found

# Create dataset with test data
test_vec <- NeuroVec(array(rnorm(1000*20), c(10,10,10,20)))
#> Error in NeuroVec(array(rnorm(1000 * 20), c(10, 10, 10, 20))): could not find function "NeuroVec"
dataset_with_test <- mvpa_dataset(train_vec, test_vec, mask=mask_vol)
#> Error: object 'train_vec' not found