Construct a design for an RSA (Representational Similarity Analysis) model

This function constructs a design for an RSA model using the provided formula, data, and optional parameters.

Usage

rsa_design(
  formula,
  data,
  block_var = NULL,
  split_by = NULL,
  keep_intra_run = FALSE
)

Arguments

formula

A formula expression specifying the dissimilarity-based regression function.

data

A named list containing the dissimilarity matrices and any other auxiliary variables.

block_var

An optional formula, character name or integer vector designating the block structure.

split_by

An optional formula indicating grouping structure for evaluating test performance.

keep_intra_run

A logical indicating whether to include within-run comparisons (default: FALSE).

Value

A list with class attributes "rsa_design" and "list", containing:

formula

The input formula

data

The input data

split_by

The split_by formula

split_groups

Grouping structure for split_by

block_var

Block structure

include

Logical vector for including/excluding comparisons

model_mat

Model matrix generated by rsa_model_mat

Details

The function creates an RSA design based on the input parameters. It checks the validity of the input data and handles splitting conditions for evaluation of test performance. It also processes optional block structures and within-run comparisons.

Examples

dismat <- dist(matrix(rnorm(100*100), 100, 100))
rdes <- rsa_design(~ dismat, list(dismat=dismat))