run_searchlight

execute a searchlight analysis

run_searchlight(model_spec, radius, method, niter, ...)

# S3 method for mvpa_model
run_searchlight(model_spec, radius = 8,
  method = c("randomized", "standard"), niter = 4, combiner = NULL,
  ...)

Arguments

model_spec
a mvpa_model instance.

radius
the searchlight radus in millimeters.

method
the type of searchlight (randomized or standard)

niter
the number of searchlight iterations (used only for 'randomized' method)

...
extra args

combiner
a function that combines results into an appropriate output.

Value

a named list of NeuroVol objects, where each element contains a performance metric (e.g. AUC) at every voxel location.

References

Bjornsdotter, M., Rylander, K., & Wessberg, J. (2011). A Monte Carlo method for locally multivariate brain mapping. Neuroimage, 56(2), 508-516.

Kriegeskorte, N., Goebel, R., & Bandettini, P. (2006). Information-based functional brain mapping. Proceedings of the National academy of Sciences of the United States of America, 103(10), 3863-3868.

Examples

dataset <- gen_sample_dataset(c(4,4,4), 100, blocks=3)
cval <- blocked_cross_validation(dataset$design$block_var)
model <- load_model("sda_notune")
mspec <- mvpa_model(model, dataset$dataset, design=dataset$design, model_type="classification", crossval=cval)
res <- run_searchlight(mspec,radius=8, method="standard")

#> INFO [2019-03-17 15:38:02] model is: sda_notune
#> INFO [2019-03-17 15:38:02] running standard searchlight with 8 radius 

#> Warning: package ‘purrr’ was built under R version 3.5.2

#> 
#> Attaching package: ‘purrr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     is_null

# a custom "combiner" can be used to post-process the output of the searchlight classifier for special cases.
# in the example below the supplied "combining function" extracts the predicted probability of the correct class 
# for every voxel and every trial and then stores them in a data.frame.

# NOT RUN {
custom_combiner <- function(mspec, good, bad) {
   good %>% pmap(function(result, id,...) {
     data.frame(trial=1:length(result$observed), id=id, prob=prob_observed(result))
   }) %>% bind_rows()
}

res2 <- run_searchlight(mspec,radius=8, method="standard", combiner=custom_combiner)
# }