select_features

feature selection on a training ROI.

Given a feature_selection specification object and a dataset return the set of selected features as a binary vector.

select_features(obj, X, Y, ...)

select_features(obj, X, Y, ...)

# S3 method for catscore
select_features(obj, X, Y, ranking.score = c("entropy",
  "avg", "max"))

# S3 method for FTest
select_features(obj, X, Y)

Arguments

obj
an object of class feature_selector

X
region of interest containing training features, a class of type ROIVolume or ROISurface or matrix

Y
the response variable

...
extra args

ranking.score
the feature score: entropy, avg, or max.

roi
the roi data, a class of inheriting from ROI class.

obj
the feature_selection object

Y
the dependent variable as a factor or numeric variable.

additional
arguments

Value

a logical vector indicating the columns of X matrix that were selected.

Examples

fsel <- feature_selector("FTest", "top_k", 2)
coords <- rbind(c(1,1,1), c(2,2,2), c(3,3,3))

ROI <- neuroim::ROIVector(neuroim::BrainSpace(c(10,10,10)), coords=coords, matrix(rnorm(100*3), 100, 3))
Y <- factor(rep(c("a", "b"), each=50))
featureMask <- select_features(fsel, ROI@data, Y)

#> selecting features via FTest

#> cutoff type top_k

#> cutoff value 2

#> retaining 2 features in matrix with 3 columns

sum(featureMask) == 2

#> [1] TRUE